Python/Line/getLineUid/requirements.txt
flask
flask_cors
flask_restful
flask_apscheduler
line-bot-sdk
loguru 
pysqlite3
gunicorn